==> finalProject/2_split_data.py <==
{"nbformat":4,"nbformat_minor":0,"metadata":{"colab":{"provenance":[],"authorship_tag":"ABX9TyMeCrthsGIGEQtadljvbRr4"},"kernelspec":{"name":"python3","display_name":"Python 3"},"language_info":{"name":"python"}},"cells":[{"cell_type":"markdown","source":["#split_data.py"],"metadata":{"id":"xKVtuiXh11II"}},{"cell_type":"code","source":["from google.colab import drive\n","drive.mount('/content/drive')"],"metadata":{"colab":{"base_uri":"https://localhost:8080/"},"id":"M8TeAZxY2h3U","executionInfo":{"status":"ok","timestamp":1749290742666,"user_tz":-540,"elapsed":17048,"user":{"displayName":"동현","userId":"02497824661603870846"}},"outputId":"de1c1695-c1a5-4bf5-f361-ad9f795083ba"},"execution_count":2,"outputs":[{"output_type":"stream","name":"stdout","text":["Mounted at /content/drive\n"]}]},{"cell_type":"code","execution_count":3,"metadata":{"colab":{"base_uri":"https://localhost:8080/"},"id":"_AAkChR_1siw","executionInfo":{"status":"ok","timestamp":1749290749215,"user_tz":-540,"elapsed":1686,"user":{"displayName":"동현","userId":"02497824661603870846"}},"outputId":"899f4f28-3fa7-4e37-af59-200e7212a311"},"outputs":[{"output_type":"stream","name":"stdout","text":["train: 55484, val: 13872\n"]}],"source":["import pandas as pd\n","from sklearn.model_selection import train_test_split\n","import os\n","\n","base_dir = '/content/drive/MyDrive/finalProject'\n","df = pd.read_csv(f'{base_dir}/hate_speech_dataset.csv')\n","train_df, val_df = train_test_split(df, test_size=0.2, random_state=42, stratify=df['label'])\n","\n","os.makedirs(f'{base_dir}/data', exist_ok=True)\n","train_df.to_csv(f'{base_dir}/data/train.csv', index=False, encoding='utf-8-sig')\n","val_df.to_csv(f'{base_dir}/data/val.csv', index=False, encoding='utf-8-sig')\n","\n","print(f\"train: {len(train_df)}, val: {len(val_df)}\")"]}]}
